Implement PReLU in a compositional way
The PReLU implementation was all over the place. This lead to a number
of bugs like https://github.com/pytorch/pytorch/issues/68760. We port it
to TI and split it into broadcasting operations (CompositeImplicit) and
kernel fusion (via TI). This makes the CPU and CUDA parts agree and
heavily simplifies the code. This also includes a number of fixes like
using opmath_t consistently and  a more efficient backward implementation
in the non-contiguous case without copying the data.

Fixes https://github.com/pytorch/pytorch/issues/68760

ghstack-source-id: f1b503c9441e5d198cff0136dc160711ae5f8651
Pull Request resolved: https://github.com/pytorch/pytorch/pull/91238