[torch.func] Add docs
Docs copy-pasted from functorch docs with minor adjustments. We are
keeping the functorch docs for BC, though that's up for debate -- we
could also just say "see .. in torch.func" for some, but not all doc
pages (we still want to keep around any examples that use
make_functional so that users can tell what the difference between that
and the new functional_call is).

Test Plan:
- docs preview

ghstack-source-id: c764403e5dc093a5f6e0f52b20ca892d22e3d54e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/91319